# Optimise count for org#index courses