--- a/app/serializers/api/v2/serializable_provider.rb
+++ b/app/serializers/api/v2/serializable_provider.rb
@@ -13,7 +13,7 @@ module API
 
       has_many :courses do
         meta do
-          { count: @object.courses.count }
+          { count: @object.courses.size }
         end
       end
     end