### Context
/organisations

### Changes proposed in this pull request
preload `courses` on `organisations`

### Guidance to review
/organisations

Before
`Completed 200 OK in 1323ms (Views: 948.1ms | ActiveRecord: 333.4ms)`

After
`Completed 200 OK in 805ms (Views: 739.2ms | ActiveRecord: 62.4ms)`